requirements.txt
streamlit==1.6
numpy==1.19.5
altair==4.1.0
pandas==1.2.5
scipy==1.6.2
click==8
docx2txt
protobuf==3.20.*
pytorch-crf
transformers
underthesea==1.3.5a3
# conda install pytorch torchvision torchaudio cudatoolkit=11.3 -c pytorch